0.3.0-beta.3:
VpWebViewScreen.kt / VpStyle.kt:
- FAB um zu Erwähnungen von Kurs zu springen mit Highlight Animation

VpScreen.kt:
- Tag/Datum zu Heute und Morgen hinzugefügt (TODO: wird nur angezeigt, wenn der Kurs Einträge hat)

- VpWebViewScreen umbenannt um Mehrdeutigkeit zu beheben

+++ app/src/main/java/de/rafaelbeckmann/hvkclient/SnackbarController.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.receiveAsFlow
 data class SnackbarEvent(
     val message: String,
     val action: SnackbarAction? = null,
-    val duration: SnackbarDuration = SnackbarDuration.Long
+    val duration: SnackbarDuration = SnackbarDuration.Short
 )
 
 data class SnackbarAction(
